asmandprog/requirements.txt
coloredlogs
pyserial
pyperclip 